requirements.txt
python-telegram-bot
fastapi~=0.111.0
uvicorn~=0.30.1
SQLAlchemy~=2.0.30
pydantic~=2.7.4
aiohttp
asyncpg
psycopg2-binary
passlib~=1.7.4
starlette~=0.37.2
python-jose